Changes to support deployment
