+++ index.html
@@ -10,6 +10,5 @@
   <body>
     <div id="root">
     </div>
-    <script src="static/bundle.js"></script>
   </body>
 </html>
