+++ package.json
@@ -5,6 +5,8 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "./node_modules/webpack/bin/webpack.js --progress --colors",
+    "watch": "node dev-server.js",
     "start": "node server.js"
   },
   "keywords": [],
@@ -23,6 +25,7 @@
     "eslint-plugin-react": "^7.0.1",
     "express": "^4.15.2",
     "extract-text-webpack-plugin": "^2.1.0",
+    "html-webpack-plugin": "^2.28.0",
     "node-sass": "^4.5.3",
     "sass-loader": "^6.0.5",
     "style-loader": "^0.18.2",
