+++ server.js
@@ -1,30 +1,11 @@
 'use strict';
 
-const path = require('path');
 const express = require('express');
-const webpack = require('webpack');
-const config = require('./webpack.config.dev');
-
 const app = express();
-const compiler = webpack(config);
+const PORT = process.env.PORT || 8080;
 
-app.use(require('webpack-dev-middleware')(compiler, {
-  noInfo: true,
-  publicPath: config.output.publicPath
-}));
+app.use(express.static(`${__dirname}/dist`));
 
-app.use(require('webpack-hot-middleware')(compiler));
-
-app.use(express.static('public'));
-
-app.get('*', function(req, res) {
-  res.sendFile(path.join(__dirname, 'index.html'));
-});
-
-app.listen(3000, 'localhost', function(err) {
-  if(err) {
-    console.log(err);
-    return;
-  }
-  console.log('Listening at http://localhost:3000');
+app.listen(PORT, function() {
+  console.log('server up:', PORT);
 });
